requirements.txt
moviepy
google-generativeai
Pillow
opencv-python
